rename reshape files to view
