+++ tests/tt_eager/tensors/test_copy_and_move.cpp
@@ -11,7 +11,7 @@
 #include "ttnn/tensor/tensor_impl.hpp"
 #include <tt-metalium/host_api.hpp>
 #include "ttnn/operations/functions.hpp"
-#include "ttnn/cpp/ttnn/operations/experimental/reshape/reshape.hpp"
+#include "ttnn/cpp/ttnn/operations/experimental/reshape/view.hpp"
 
 using namespace tt;
 using namespace tt_metal;
