+++ ttnn/cpp/ttnn/operations/data_movement/fold/fold.cpp
@@ -14,7 +14,7 @@
 #include "cpp/ttnn/operations/data_movement/pad/pad.hpp"
 #include <tt-metalium/constants.hpp>
 
-#include "ttnn/cpp/ttnn/operations/experimental/reshape/reshape.hpp"
+#include "ttnn/cpp/ttnn/operations/experimental/reshape/view.hpp"
 
 #include "fold.hpp"
